added bottom navbar and pages and changed to posts instead of students
