+++ app/src/main/java/com/example/mobile_application_course/MainActivity.kt
@@ -4,6 +4,7 @@ package com.example.mobile_application_course
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
+import android.widget.TextView
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
@@ -12,6 +13,7 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI
+import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
 
@@ -28,7 +30,11 @@ class MainActivity : AppCompatActivity() {
         }
 
         val toolBar: Toolbar = findViewById(R.id.main_toolbar)
+        val toolbarTitle: TextView = findViewById(R.id.toolbar_title)
         setSupportActionBar(toolBar)
+
+        supportActionBar?.setDisplayShowTitleEnabled(false)
+        toolbarTitle.text = getString(R.string.app_name)
 
         val navHostFragment: NavHostFragment? =
             supportFragmentManager.findFragmentById(R.id.main_nav_host) as? NavHostFragment
@@ -39,6 +45,13 @@ class MainActivity : AppCompatActivity() {
                 navController = it,
             )
         }
+
+        val bottomNavigationView: BottomNavigationView = findViewById(R.id.main_bottom_nav)
+        navController?.let { NavigationUI.setupWithNavController(bottomNavigationView, it) }
+
+        navController?.let {it.addOnDestinationChangedListener { _, destination, _ ->
+            supportActionBar?.setDisplayHomeAsUpEnabled(false)
+        }}
     }
 
 
@@ -46,25 +59,25 @@ class MainActivity : AppCompatActivity() {
         return super.onCreateOptionsMenu(menu)
     }
 
-    override fun onPrepareOptionsMenu(menu: Menu?): Boolean {
-        val currentDestination = navController?.currentDestination
-
-        when (currentDestination?.id) {
-            R.id.studentsListFragment -> {
-                menuInflater.inflate(R.menu.menu_new_student, menu)
-            }
-
-            R.id.studentDetailsFragment -> {
-                menuInflater.inflate(R.menu.menu_edit_student, menu)
-            }
-
-            else -> {
-                menu?.clear()
-            }
-        }
-
-        return super.onPrepareOptionsMenu(menu)
-    }
+//    override fun onPrepareOptionsMenu(menu: Menu?): Boolean {
+//        val currentDestination = navController?.currentDestination
+//
+//        when (currentDestination?.id) {
+//            R.id.homeFragment -> {
+//                menuInflater.inflate(R.menu.menu_new_student, menu)
+//            }
+//
+//            R.id.postDetailsFragment -> {
+//                menuInflater.inflate(R.menu.menu_edit_student, menu)
+//            }
+//
+//            else -> {
+//                menu?.clear()
+//            }
+//        }
+//
+//        return super.onPrepareOptionsMenu(menu)
+//    }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
@@ -73,12 +86,12 @@ class MainActivity : AppCompatActivity() {
                 true
             }
 
-            R.id.newStudentFragment -> {
-                val action =
-                    StudentsListFragmentDirections.actionGlobalNewStudentFragment()
-                navController?.navigate(action)
-                true
-            }
+//            R.id.newPostFragment -> {
+//                val action =
+//                    HomeFragmentDirections.actionGlobalNewStudentFragment()
+//                navController?.navigate(action)
+//                true
+//            }
 
             else -> super.onOptionsItemSelected(item)
         }
