+++ app/src/main/res/layout/fragment_new_post.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditStudentFragment">
+    tools:context=".NewPostFragment">
 
     <include
         android:id="@+id/student_input_form"
         layout="@layout/student_input_form" />
 
     <Button
-        android:id="@+id/edit_student_fragment_cancel_button"
+        android:id="@+id/add_student_fragment_cancel_button"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
@@ -20,34 +20,21 @@
         android:backgroundTint="#2196F3"
         android:text="Cancel"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/edit_student_fragment_delete_button"
-        app:layout_constraintHorizontal_bias="0.5" />
-
-    <Button
-        android:id="@+id/edit_student_fragment_delete_button"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="4dp"
-        app:layout_constraintBaseline_toBaselineOf="@+id/edit_student_fragment_cancel_button"
-        android:backgroundTint="#2196F3"
-        android:text="Delete"
-        app:layout_constraintEnd_toStartOf="@+id/edit_student_fragment_save_button"
+        app:layout_constraintEnd_toStartOf="@+id/add_student_fragment_save_button"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/edit_student_fragment_cancel_button" />
+        app:layout_constraintStart_toStartOf="parent" />
 
     <Button
-        android:id="@+id/edit_student_fragment_save_button"
+        android:id="@+id/add_student_fragment_save_button"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="4dp"
         android:layout_marginEnd="8dp"
-        app:layout_constraintBaseline_toBaselineOf="@+id/edit_student_fragment_delete_button"
         android:backgroundTint="#2196F3"
         android:text="Save"
+        app:layout_constraintBaseline_toBaselineOf="@+id/add_student_fragment_cancel_button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/edit_student_fragment_delete_button" />
+        app:layout_constraintStart_toEndOf="@+id/add_student_fragment_cancel_button" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
