+++ app/src/main/res/layout/fragment_post_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudentDetailsFragment">
+    tools:context=".PostDetailsFragment">
 
     <ImageView
         android:id="@+id/student_details_fragment_image_view"
