+++ app/src/main/res/layout/fragment_posts_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudentsListFragment">
+    tools:context=".HomeFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
