+++ app/src/main/res/menu/menu_edit_student.xml
@@ -2,7 +2,7 @@
 <menu xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/editStudentFragment"
+        android:id="@+id/editPostFragment"
         app:showAsAction="ifRoom|withText"
         android:title="Edit" />
 </menu>
