+++ app/src/main/res/menu/menu_new_student.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/newStudentFragment"
+        android:id="@+id/newPostFragment"
         android:icon="@drawable/baseline_add_24"
         android:title="Add Student"
         app:showAsAction="ifRoom" />
