+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,44 +3,66 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/studentsListFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
-        android:id="@+id/studentsListFragment"
-        android:name="com.example.mobile_application_course.StudentsListFragment"
-        android:label="Students List"
-        tools:layout="@layout/fragment_students_list">
+        android:id="@+id/homeFragment"
+        android:name="com.example.mobile_application_course.HomeFragment"
+        tools:layout="@layout/fragment_posts_list">
         <action
             android:id="@+id/action_studentsListFragment_to_studentDetailsFragment"
-            app:destination="@id/studentDetailsFragment" />
+            app:destination="@id/postDetailsFragment" />
     </fragment>
-    <fragment
-        android:id="@+id/newStudentFragment"
-        android:name="com.example.mobile_application_course.NewStudentFragment"
-        android:label="Add Student"
-        tools:layout="@layout/fragment_new_student" />
+
     <action
         android:id="@+id/action_global_newStudentFragment"
-        app:destination="@id/newStudentFragment" />
+        app:destination="@id/newPostFragment" />
+
     <fragment
-        android:id="@+id/studentDetailsFragment"
-        android:name="com.example.mobile_application_course.StudentDetailsFragment"
+        android:id="@+id/postDetailsFragment"
+        android:name="com.example.mobile_application_course.PostDetailsFragment"
         android:label="Student Details"
-        tools:layout="@layout/fragment_student_details">
+        tools:layout="@layout/fragment_post_details">
         <argument
             android:name="position"
             app:argType="integer" />
         <action
             android:id="@+id/action_studentDetailsFragment_to_editStudentFragment"
-            app:destination="@id/editStudentFragment" />
+            app:destination="@id/editPostFragment" />
     </fragment>
+
     <fragment
-        android:id="@+id/editStudentFragment"
-        android:name="com.example.mobile_application_course.EditStudentFragment"
+        android:id="@+id/editPostFragment"
+        android:name="com.example.mobile_application_course.EditPostFragment"
         android:label="Edit Student"
-        tools:layout="@layout/fragment_edit_student">
+        tools:layout="@layout/fragment_edit_post">
         <argument
             android:name="position"
             app:argType="integer" />
     </fragment>
+
+    <fragment
+        android:id="@+id/newPostFragment"
+        android:name="com.example.mobile_application_course.NewPostFragment"
+        android:label="Add Student"
+        tools:layout="@layout/fragment_new_post" />
+
+    <fragment
+        android:id="@+id/searchFragment"
+        android:name="com.example.mobile_application_course.SearchFragment"
+        android:label="Search"
+        tools:layout="@layout/fragment_search"/>
+
+    <fragment
+        android:id="@+id/profileFragment"
+        android:name="com.example.mobile_application_course.ProfileFragment"
+        android:label="Search"
+        tools:layout="@layout/fragment_profile"/>
+
+    <fragment
+        android:id="@+id/savedPostsFragment"
+        android:name="com.example.mobile_application_course.SavedPostsFragment"
+        android:label="Search"
+        tools:layout="@layout/fragment_saved_posts"/>
+
 </navigation>
