+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Mobile-Application-Course</string>
+    <string name="app_name">Foodie Finder</string>
 </resources>
